update OcticonComponent to correctly deal with test_selector
